refactor mapper of trusted apps for tags-to-effectScope from server code to common
